[Documentation]
- Renamed some files

- Added doc foundations.
