--- GodtierGeometryScriptUtils/GodtierGeometryScriptUtils.uplugin
@@ -1,29 +1,0 @@
-﻿{
-	"FileVersion": 3,
-	"Version": 1,
-	"VersionName": "1.0",
-	"FriendlyName": "GodtierGeometryScriptUtils",
-	"Description": "An array of helper tools to make geometry script similar to Houdini",
-	"Category": "Other",
-	"CreatedBy": "Greggory Addison",
-	"CreatedByURL": "www.godtiergames.net",
-	"DocsURL": "",
-	"MarketplaceURL": "",
-	"CanContainContent": true,
-	"IsBetaVersion": false,
-	"IsExperimentalVersion": false,
-	"Installed": false,
-	"Modules": [
-		{
-			"Name": "GodtierGeometryScriptUtils",
-			"Type": "Runtime",
-			"LoadingPhase": "PreLoadingScreen"
-		}
-	],
-	"Plugins": [
-		{
-			"Name": "GeometryScripting",
-			"Enabled": true
-		}
-	]
-}
