--- GodtierGeometryScriptUtils/Source/GodtierGeometryScriptUtils/GodtierGeometryScriptUtils.Build.cs
@@ -1,73 +1,0 @@
-﻿// Some copyright should be here...
-
-using UnrealBuildTool;
-
-public class GodtierGeometryScriptUtils : ModuleRules
-{
-	public GodtierGeometryScriptUtils(ReadOnlyTargetRules Target) : base(Target)
-	{
-		PCHUsage = ModuleRules.PCHUsageMode.UseExplicitOrSharedPCHs;
-		
-		PublicIncludePaths.AddRange(
-			new string[] {
-				// ... add public include paths required here ...
-			}
-			);
-				
-		
-		PrivateIncludePaths.AddRange(
-			new string[] {
-				// ... add other private include paths required here ...
-			}
-			);
-			
-		
-		PublicDependencyModuleNames.AddRange(
-			new string[]
-			{
-				"Core",
-				"GeometryFramework",
-				"GeometryScriptingCore",
-				"GeometryCore",
-				"ModelingOperators", 
-				"MeshModelingTools",
-				"DynamicMesh",
-				"ModelingComponents",
-				"ModelingOperators",
-
-
-				// ... add other public dependencies that you statically link with here ...
-			}
-			);
-			
-		
-		PrivateDependencyModuleNames.AddRange(
-			new string[]
-			{
-				"CoreUObject",
-				"Engine",
-				"Slate",
-				"SlateCore",
-				"GeometryFramework",
-				"GeometryScriptingCore",
-				"GeometryCore", 
-				"ModelingOperators",
-				"MeshModelingTools",
-				"DynamicMesh",
-				"ModelingComponents",
-				"ModelingOperators",
-
-
-				// ... add private dependencies that you statically link with here ...	
-			}
-			);
-		
-		
-		DynamicallyLoadedModuleNames.AddRange(
-			new string[]
-			{
-				// ... add any modules that your module loads dynamically here ...
-			}
-			);
-	}
-}
