--- GodtierGeometryScriptUtils/Source/GodtierGeometryScriptUtils/Private/GodtierGeometryScriptUtils.cpp
@@ -1,22 +1,0 @@
-﻿// Copyright Epic Games, Inc. All Rights Reserved.
-
-#include "GodtierGeometryScriptUtils.h"
-
-#define LOCTEXT_NAMESPACE "FGodtierGeometryScriptUtilsModule"
-
-void FGodtierGeometryScriptUtilsModule::StartupModule()
-{
-	// This code will execute after your module is loaded into memory; the exact timing is specified in the .uplugin file per-module
-	
-}
-
-void FGodtierGeometryScriptUtilsModule::ShutdownModule()
-{
-	// This function may be called during shutdown to clean up your module.  For modules that support dynamic reloading,
-	// we call this function before unloading the module.
-	
-}
-
-#undef LOCTEXT_NAMESPACE
-	
-IMPLEMENT_MODULE(FGodtierGeometryScriptUtilsModule, GodtierGeometryScriptUtils)
