--- GodtierGeometryScriptUtils/Source/GodtierGeometryScriptUtils/Private/SplineUtilities/GodtierSplineUtilities.cpp
@@ -1,4 +1,0 @@
-﻿// Fill out your copyright notice in the Description page of Project Settings.
-
-
-#include "SplineUtilities/GodtierSplineUtilities.h"
