--- GodtierGeometryScriptUtils/Source/GodtierGeometryScriptUtils/Public/GodtierGeometryScriptUtils.h
@@ -1,14 +1,0 @@
-﻿// Copyright Epic Games, Inc. All Rights Reserved.
-
-#pragma once
-
-#include "Modules/ModuleManager.h"
-
-class FGodtierGeometryScriptUtilsModule : public IModuleInterface
-{
-public:
-
-	/** IModuleInterface implementation */
-	virtual void StartupModule() override;
-	virtual void ShutdownModule() override;
-};
