--- GodtierGeometryScriptUtils/Source/GodtierGeometryScriptUtils/Public/SplineUtilities/GodtierSplineUtilities.h
@@ -1,17 +1,0 @@
-﻿// Fill out your copyright notice in the Description page of Project Settings.
-
-#pragma once
-
-#include "CoreMinimal.h"
-#include "Kismet/BlueprintFunctionLibrary.h"
-#include "GodtierSplineUtilities.generated.h"
-
-/**
- * 
- */
-UCLASS()
-class GODTIERGEOMETRYSCRIPTUTILS_API UGodtierSplineUtilities : public UBlueprintFunctionLibrary
-{
-	GENERATED_BODY()
-	
-};
