+++ HandyMan/HandyMan.uplugin
@@ -78,7 +78,7 @@
 			"Enabled": true
 		},
 		{
-			"Name": "GodtierGeometryScriptUtils",
+			"Name": "HandyManGeometryScriptUtils",
 			"Enabled": true
 		},
 		{
