+++ HandyMan/Source/HandyMan/HandyMan.Build.cs
@@ -63,7 +63,7 @@ public class HandyMan : ModuleRules
 				"ModelingComponents",
 				"Landscape",
 				"MeshModelingTools",
-				"GodtierGeometryScriptUtils",
+				"HandyManGeometryScriptUtils",
 				"GeometryScriptingEditor",
 				"DynamicMesh", 
 				"MeshModelingToolsExp",
@@ -125,7 +125,7 @@ public class HandyMan : ModuleRules
 				"PCGWaterInterop",
 				"PCGGeometryScriptInterop", 
 				"EditorScriptingUtilities", 
-				"GodtierGeometryScriptUtils",
+				"HandyManGeometryScriptUtils",
 				"GeometryScriptingEditor",
 				"GeometryScriptingCore",
 				"EditorStyle",
