+++ HandyManGeometryScriptUtils/Source/HandyManGeometryScriptUtils/Private/HandyManGeometryScriptUtils.cpp
@@ -1,0 +1,22 @@
+﻿// Copyright Epic Games, Inc. All Rights Reserved.
+
+#include "HandyManGeometryScriptUtils.h"
+
+#define LOCTEXT_NAMESPACE "FHandyManGeometryScriptUtilsModule"
+
+void FHandyManGeometryScriptUtilsModule::StartupModule()
+{
+	// This code will execute after your module is loaded into memory; the exact timing is specified in the .uplugin file per-module
+	
+}
+
+void FHandyManGeometryScriptUtilsModule::ShutdownModule()
+{
+	// This function may be called during shutdown to clean up your module.  For modules that support dynamic reloading,
+	// we call this function before unloading the module.
+	
+}
+
+#undef LOCTEXT_NAMESPACE
+	
+IMPLEMENT_MODULE(FHandyManGeometryScriptUtilsModule, HandyManGeometryScriptUtils)
