+++ HandyManGeometryScriptUtils/Source/HandyManGeometryScriptUtils/Public/HandyManGeometryScriptUtils.h
@@ -1,0 +1,14 @@
+﻿// Copyright Epic Games, Inc. All Rights Reserved.
+
+#pragma once
+
+#include "Modules/ModuleManager.h"
+
+class FHandyManGeometryScriptUtilsModule : public IModuleInterface
+{
+public:
+
+	/** IModuleInterface implementation */
+	virtual void StartupModule() override;
+	virtual void ShutdownModule() override;
+};
