+++ HandyManGeometryScriptUtils/Source/HandyManGeometryScriptUtils/Public/SplineUtilities/HandyManSplineUtilities.h
@@ -1,0 +1,17 @@
+﻿// Fill out your copyright notice in the Description page of Project Settings.
+
+#pragma once
+
+#include "CoreMinimal.h"
+#include "Kismet/BlueprintFunctionLibrary.h"
+#include "HandyManSplineUtilities.generated.h"
+
+/**
+ * 
+ */
+UCLASS()
+class HANDYMANGEOMETRYSCRIPTUTILS_API UHandyManSplineUtilities : public UBlueprintFunctionLibrary
+{
+	GENERATED_BODY()
+	
+};
