+++ doc/tools/world/BuildingGenerator.md
@@ -1,0 +1,1 @@
+﻿
